requirements.txt
pyrofork
uvloop
motor
cachetools
telethon
requests
speedtest-cli
beautifulsoup4
pytz
unidecode
telegraph
pillow
httpx[http2]
python-arq
gtts
gpytranslate
opencv-python
blackpink
openai
tgcrypto
apscheduler
lexica-api
python-telegram-bot
yt-dlp
moviepy
psutil
httpx==0.27.2
rich
aiofiles
aiohttp==3.9.5
asyncio
dnspython
ffmpeg-python
numpy
youtube-search
youtube-search-python
pyjokes
lyricsgenius
